pyproject.toml
[project]
name = "nonebot-plugin-cnrail"
dynamic = ["version"]
description = "NoneBot2 plugin for query the train time table"
authors = [
    { name = "student_2333", email = "lgc2333@126.com" },
    { name = "XieXiLin", email = "support@xiexilin.com" },
]
dependencies = [
    "nonebot2>=2.3.1",
    "httpx>=0.27.0",
    "nonebot-plugin-alconna>=0.48.0",
    "nonebot-plugin-htmlrender>=0.3.2",
    "jinja2>=3.1.4",
    "pytz>=2024.1",
    "cookit[jinja,loguru,playwright,pyd]>=0.8.1",
]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-cnrail"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_cnrail/__init__.py"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
